backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-keycloak
python-dotenv
itsdangerous
